+++ src/ts/ir/process.ts
@@ -53,8 +53,8 @@ export const processAfterRender = (vditor: IVditor, options = {
             vditor.options.input(text);
         }
 
-        if (vditor.options.counter !== false) {
-            vditor.counter.render(text, vditor.options.counter);
+        if (vditor.options.counter.enable) {
+            vditor.counter.render(vditor, text);
         }
 
         if (vditor.options.cache.enable) {